+++ .gitignore
@@ -1,0 +1,16 @@
+__pycache__/
+*.pyc
+*.pyo
+.env
+.venv/
+venv/
+.envrc
+.DS_Store
+pytest_cache/
+.coverage
+.python-version
+.env/
+
+# Virtual environment folders
+.vscode/
+.idea/
