+++ modified_tracking_budget/__init__.py
@@ -1,0 +1,10 @@
+# Package initializer for modified_tracking_budget
+__all__ = [
+    "budget_manager",
+    "database",
+    "main",
+    "temporary",
+]
+
+# Expose top-level package version (update as needed)
+__version__ = "0.0.0"
